requirements.txt
numpy
scipy
scikit-image
libigl
polyscope
tensorboard
requests
tqdm
bpy
